designpatterns

+++ src/main/java/com/thoughtfocus/behavioraldesignpatterns/chainofresponsibility/Chain.java
@@ -2,6 +2,6 @@ package com.thoughtfocus.behavioraldesignpatterns.chainofresponsibility;
 
 public interface Chain {
     public void setNextChain(Chain nextChain);
-    public short calculate(Numbers request);
+    public int calculate(Numbers request);
 
 }
